+++ core/src/main/scala/core/DataFrameFromTo.scala
@@ -24,8 +24,6 @@ import java.time.Instant
 import java.util
 import java.util.{Calendar, Properties, UUID}
 
-import javax.mail.internet.{InternetAddress, MimeMessage}
-import javax.mail.{Message, Session, Transport}
 import com.amazonaws.services.logs.model.{DescribeLogStreamsRequest, InputLogEvent, PutLogEventsRequest}
 import com.amazonaws.services.s3.model._
 import com.amazonaws.services.s3.{AmazonS3, AmazonS3ClientBuilder}
@@ -41,12 +39,12 @@ import com.mongodb.{MongoClient, MongoClientURI}
 import config.AppConfig
 import core.DataPull.jsonObjectPropertiesToMap
 import helper._
-import org.apache.avro.Schema
-import org.apache.avro.generic.{GenericData, GenericRecord}
+import javax.mail.internet.{InternetAddress, MimeMessage}
+import javax.mail.{Message, Session, Transport}
+import org.apache.avro.generic.GenericData
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FSDataOutputStream, FileSystem, Path}
 import org.apache.kafka.clients.consumer.KafkaConsumer
-import org.apache.kafka.clients.producer.{KafkaProducer, ProducerConfig, ProducerRecord}
 import org.apache.spark.SparkConf
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.functions.{col, struct}
@@ -58,7 +56,6 @@ import org.codehaus.jettison.json.JSONObject
 import org.elasticsearch.spark.sql._
 import org.influxdb.InfluxDBFactory
 import org.influxdb.dto.Query
-import org.json.simple.parser.JSONParser
 import security._
 import za.co.absa.abris.avro.functions.to_confluent_avro
 import za.co.absa.abris.avro.read.confluent.SchemaManager
@@ -95,8 +92,8 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline: String) extends Serializab
     var filePrefix = ""
 
     if (isS3) {
-      filePrefix = "s3a://"
-      sparkSession.conf.set("fs.s3a.connection.maximum", 100)
+      filePrefix = "s3://"
+      sparkSession.conf.set("fs.s3.connection.maximum", 100)
     }
 
     def createOrReplaceTempViewOnDF(df: org.apache.spark.sql.DataFrame): org.apache.spark.sql.DataFrame = {
@@ -128,6 +125,8 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline: String) extends Serializab
 
       } else if (fileFormat == "avro") {
         createOrReplaceTempViewOnDF(sparkSession.read.format("avro").load(s"$filePrefix$filePath"))
+      } else if (fileFormat == "orc") {
+        createOrReplaceTempViewOnDF(sparkSession.read.format("orc").load(s"$filePrefix$filePath"))
       } else {
         //parquet
         createOrReplaceTempViewOnDF(sparkSession.read.option("mergeSchema", mergeSchema).parquet(s"$filePrefix$filePath"))
@@ -255,8 +254,8 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline: String) extends Serializab
     }
 
     if (isS3) {
-      filePrefix = "s3a://"
-      sparkSession.conf.set("fs.s3a.connection.maximum", 100)
+      filePrefix = "s3://"
+      sparkSession.conf.set("fs.s3.connection.maximum", 100)
     }
     if (isSFTP) {
 
@@ -335,6 +334,21 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline: String) extends Serializab
             .option("header", "true")
             .mode(SaveMode.valueOf(s3SaveMode))
             .save(s"$filePrefix$filePath")
+        }
+      } else if (fileFormat == "orc") {
+        if (groupByFields == "") {
+          dft
+            .write
+            .option("header", "true")
+            .mode(SaveMode.valueOf(s3SaveMode))
+            .orc(s"$filePrefix$filePath")
+        } else {
+          dft
+            .write
+            .partitionBy(groupByFieldsArray: _*)
+            .option("header", "true")
+            .mode(SaveMode.valueOf(s3SaveMode))
+            .orc(s"$filePrefix$filePath")
         }
       }
       else {
@@ -1117,6 +1131,9 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline: String) extends Serializab
                         topic: String,
                         keyField: String,
                         keyFormat: String,
+                        addlSparkOptions: JSONObject,
+                        valueField: String,
+                        headerField: String,
                         df: org.apache.spark.sql.DataFrame): Unit = {
     val commonRegistryConfig = Map(
       SchemaManager.PARAM_SCHEMA_REGISTRY_TOPIC -> topic,
@@ -1131,15 +1148,35 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline: String) extends Serializab
       //, SchemaManager.PARAM_VALUE_SCHEMA_ID -> "latest"
     )
 
-    val allColumns = struct(df.columns.head, df.columns.tail: _*)
+    var allColumns = struct(df.columns.head, df.columns.tail: _*)
+
+    if (valueField != null) {
+      allColumns = df.col(valueField)
+    }
+
+    val isHeaderPresent = df.columns.contains("headers")
+
+    if (headerField != null) {
+      df.withColumn("headers", df.col(headerField))
+
+    } else if (!isHeaderPresent) {
+      import org.apache.spark.sql.functions.{lit, struct}
+      df.withColumn("headers", struct(lit(System.currentTimeMillis()) as 'time))
+    }
+
+    var sparkOptions = Map("kafka.bootstrap.servers" -> bootstrapServers, "includeHeaders" -> "true")
+
+    if (addlSparkOptions != null) {
+      sparkOptions = sparkOptions ++ jsonObjectPropertiesToMap(addlSparkOptions)
+    }
     try {
       df.select((keyFormat match {
         case "avro" => to_confluent_avro(col(keyField), keyRegistryConfig)
         case _ => col(keyField).cast(StringType)
       })
-        as 'key, to_confluent_avro(allColumns, valueRegistryConfig) as 'value)
+        as 'key, to_confluent_avro(allColumns, valueRegistryConfig) as 'value, df.col("headers"))
         .write
-        .option("kafka.bootstrap.servers", bootstrapServers)
+        .options(sparkOptions)
         .option("topic", topic)
         .format("kafka")
         .save()
@@ -1157,7 +1194,7 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline: String) extends Serializab
     if (isWindowsAuthenticated.toBoolean) {
       if (platform == "mssql") {
         driver = "net.sourceforge.jtds.jdbc.Driver"
-        url = "jdbc:jtds:sqlserver://" + server + ":" + (if (port == null) "1433" else port) + "/" + database + ";domain= " + domainName + ";useNTLMv2=true"
+        url = "jdbc:jtds:sqlserver://" + server + ":" + (if (port == null) "1433" else port) + "/" + database + ";domain=" + domainName + ";useNTLMv2=true"
       }
     } else {
       if (platform == "mssql") {